+++ app/routes/pos.py
